requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
chatgpt-block==0.1.6
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.0
pydantic==2.9.2
pydantic-core==2.23.4
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
uagents
groq
flask>=2.0
flask_cors
python-dotenv
google-generativeai>=0.4.0 